Fix compilation warning in bio_end_erase

+++ dm-secdel.c
@@ -127,8 +127,8 @@ static void bio_end_erase(struct bio *bio)
 	int i;
 
 	if (bio->bi_status)
-		DMERR("bio_end_erase %lu[%u] error=%d",
-		    bio->bi_iter.bi_sector,
+		DMERR("bio_end_erase %llu[%u] error=%d",
+		    (long long unsigned int)bio->bi_iter.bi_sector,
 		    bio->bi_iter.bi_size >> 9,
 		    bio->bi_status);
 	bio_for_each_segment_all(bvec, bio, i)
